requirements.txt
openai
python-dotenv
pyodbc